VIITE-568 refactor to separate object class
